renamed provisioning wizard to connection wizard to better reflect scope
