+++ app/src/main/res/layout/activity_connection_wizard_page1.xml
