+++ app/src/main/res/layout/activity_connection_wizard_page2.xml
